split some processors into different files
